+++ src/CareTogether.Core/Engines/PolicyEvaluation/ApprovalCalculations.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 
-namespace CareTogether.Engines
+namespace CareTogether.Engines.PolicyEvaluation
 {
     internal static class ApprovalCalculations
     {
