+++ src/CareTogether.Core/Engines/PolicyEvaluation/PolicyEvaluationEngine.cs
@@ -1,0 +1,46 @@
+using CareTogether.Resources;
+using System;
+using System.Collections.Immutable;
+using System.Threading.Tasks;
+
+namespace CareTogether.Engines.PolicyEvaluation
+{
+    public sealed class PolicyEvaluationEngine : IPolicyEvaluationEngine
+    {
+        private readonly IPoliciesResource policiesResource;
+
+
+        public PolicyEvaluationEngine(IPoliciesResource policiesResource)
+        {
+            this.policiesResource = policiesResource;
+        }
+
+
+        public async Task<VolunteerFamilyApprovalStatus> CalculateVolunteerFamilyApprovalStatusAsync(Guid organizationId, Guid locationId,
+            Family family, ImmutableList<CompletedRequirementInfo> completedFamilyRequirements,
+            ImmutableList<ExemptedRequirementInfo> exemptedFamilyRequirements,
+            ImmutableList<RemovedRole> removedFamilyRoles,
+            ImmutableDictionary<Guid, ImmutableList<CompletedRequirementInfo>> completedIndividualRequirements,
+            ImmutableDictionary<Guid, ImmutableList<ExemptedRequirementInfo>> exemptedIndividualRequirements,
+            ImmutableDictionary<Guid, ImmutableList<RemovedRole>> removedIndividualRoles)
+        {
+            var policy = await policiesResource.GetCurrentPolicy(organizationId, locationId);
+
+            return ApprovalCalculations.CalculateVolunteerFamilyApprovalStatus(
+                policy.VolunteerPolicy, family, DateTime.UtcNow,
+                completedFamilyRequirements, exemptedFamilyRequirements, removedFamilyRoles,
+                completedIndividualRequirements, exemptedIndividualRequirements, removedIndividualRoles);
+        }
+
+
+        public async Task<ReferralStatus> CalculateReferralStatusAsync(
+            Guid organizationId, Guid locationId, Family family,
+            ReferralEntry referralEntry)
+        {
+            var policy = await policiesResource.GetCurrentPolicy(organizationId, locationId);
+
+            return ReferralCalculations.CalculateReferralStatus(
+                policy.ReferralPolicy, referralEntry, DateTime.UtcNow);
+        }
+    }
+}
