+++ src/CareTogether.Core/Managers/CombinedFamilyInfoFormatter.cs
@@ -1,7 +1,8 @@
-﻿using CareTogether.Engines;
+﻿using CareTogether.Engines.Authorization;
+using CareTogether.Engines.PolicyEvaluation;
 using CareTogether.Resources;
-using System;
 using Nito.AsyncEx;
+using System;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Security.Claims;
