+++ src/CareTogether.Core/Managers/DirectoryManager.cs
@@ -1,4 +1,4 @@
-﻿using CareTogether.Engines;
+﻿using CareTogether.Engines.Authorization;
 using CareTogether.Resources;
 using Nito.AsyncEx;
 using System;
