+++ src/CareTogether.Core/Managers/ReferralManager.cs
@@ -1,4 +1,4 @@
-﻿using CareTogether.Engines;
+﻿using CareTogether.Engines.Authorization;
 using CareTogether.Resources;
 using System;
 using System.Security.Claims;
