+++ src/CareTogether.Core/Managers/SharedContracts.cs
@@ -1,4 +1,5 @@
 ﻿using CareTogether.Engines;
+using CareTogether.Engines.PolicyEvaluation;
 using CareTogether.Resources;
 using System;
 using System.Collections.Immutable;
