+++ test/CareTogether.Core.Test/ApprovalCalculationTests/FamilyRequirementMissingForIndividualsTests.cs
@@ -1,4 +1,5 @@
 ﻿using CareTogether.Engines;
+using CareTogether.Engines.PolicyEvaluation;
 using CareTogether.Resources;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
