+++ test/CareTogether.Core.Test/AssertEx.cs
@@ -1,4 +1,5 @@
 ﻿using CareTogether.Engines;
+using CareTogether.Engines.PolicyEvaluation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
