+++ test/CareTogether.Core.Test/ReferralCalculationTests/CalculateMissingSetupRequirements.cs
@@ -1,4 +1,5 @@
 ﻿using CareTogether.Engines;
+using CareTogether.Engines.PolicyEvaluation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace CareTogether.Core.Test.ReferralCalculationTests
